requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
numpy
scikit-learn
xgboost
joblib
openpyxl
shap
passlib[bcrypt]
bcrypt==4.0.1
python-jose
authlib
python-multipart
python-dotenv
email-validator
requests
fastapi-mail
pandas
rapidfuzz
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
gunicorn
python-dateutil